requirements.txt
opencv-python
tensorflow
pandas
tensorflow-hub
pymongo
pyfcm==2.0.2
numpy
tqdm
boto3
tqdm
python-dotenv
